@@ -1,7 +1,11 @@
-const geckos = require('../../packages/server/lib').default
-const http = require('http')
-const express = require('express')
-const path = require('path')
+/* eslint-disable sort-imports */
+import express  from 'express'
+import geckos from '../../packages/server/lib/index.js'
+import http  from 'http'
+import path from 'path'
+
+import {__dirname} from './_dirname.js'
+
 const app = express()
 const server = http.createServer(app)
 const io = geckos()
@@ -41,6 +45,15 @@ describe('connection', () => {
 page.goto('http://localhost:5100/e2e/emptyMessages.html')
 
 afterAll(async () => {
-  page.close()
-  server.close()
+  const close = () => {
+    return new Promise(resolve => {
+      server.close(() => {
+        resolve()
+      })
+    })
+  }
+
+  await close()
+  // await page.close()
+  // await browser.close()
 })
