@@ -1,8 +1,12 @@
-const geckos = require('../../packages/server/lib').default
-const io = geckos()
+/* eslint-disable sort-imports */
+import express  from 'express'
+import geckos from '../../packages/server/lib/index.js'
+import http  from 'http'
+import path from 'path'
 
-const express = require('express')
-const path = require('path')
+import {__dirname} from './_dirname.js'
+
+const io = geckos()
 const app = express()
 
 app.use('/', express.static(path.join(__dirname, '../')))
@@ -43,6 +47,6 @@ describe('connection', () => {
 
 page.goto('http://localhost:5600/e2e/forwardMessage.html')
 
-afterAll(async () => {
-  page.close()
-})
+// afterAll(async () => {
+//   page.close()
+// })
