@@ -1,23 +1,27 @@
-import EventEmitter from 'eventemitter3'
-import { Data, EventOptions, ChannelId, EventName, RoomId } from './types'
+import { ChannelId, Data, EventName, EventOptions, RoomId } from './types.js'
+import { Events } from '@yandeu/events'
+
+interface BridgeEventMap {
+  [key: string]: (data?: Data, options?: EventOptions) => void
+}
 
 export class Bridge {
-  eventEmitter = new EventEmitter()
+  eventEmitter = new Events<BridgeEventMap>()
 
   emit(
     eventName: EventName,
     data?: Data,
     connection: {
-      senderId?: ChannelId
       id?: ChannelId
       roomId?: RoomId
+      senderId?: ChannelId
     } = {}
   ) {
     this.eventEmitter.emit(eventName, data, connection)
   }
 
   on(eventName: EventName, cb: Function) {
-    return this.eventEmitter.on(eventName, (data: Data, options: EventOptions) => {
+    return this.eventEmitter.on(eventName, (data, options) => {
       cb(data, options)
     })
   }
