@@ -1,14 +1,22 @@
-import { ArrayBufferView, RawMessage, Data } from './types'
+import { ArrayBufferView, Data, RawMessage } from './types.js'
 
-const isRawMessage = (data: Data | RawMessage) => {
-  return typeof data === 'string' || data instanceof ArrayBuffer || data instanceof ArrayBufferView
+// const isRawMessage = (data: Data | RawMessage) => {
+//   return typeof data === 'string' || isBufferMessage(data)
+// }
+
+const isStringMessage = (data: any) => {
+  return typeof data === 'string'
+}
+
+const isBufferMessage = (data: any) => {
+  return data instanceof ArrayBuffer || data instanceof ArrayBufferView
 }
 
 const isObject = (data: Data) => {
   return typeof data === 'object'
 }
 
-const isJSONString = (data: Data) => {
+const isJSONMessage = (data: Data) => {
   try {
     // check if it is a string
     if (typeof data !== 'string') return false
@@ -22,4 +30,4 @@ const isJSONString = (data: Data) => {
   }
 }
 
-export { isRawMessage, isObject, isJSONString }
+export { isStringMessage, isBufferMessage, isObject, isJSONMessage }
