@@ -1,25 +1,44 @@
-import { EVENTS, ERRORS } from './constants'
-import { isRawMessage, isJSONString } from './helpers'
+import { ERRORS, EVENTS } from './constants.js'
+import { isBufferMessage, isJSONMessage, isStringMessage } from './helpers.js'
 
 const ParseMessage = (ev: MessageEvent) => {
   let { data } = ev
-  let key
-  let parsedData
-  let JSONString = isJSONString(data)
+  if (!data) data = ev
 
-  if (!JSONString && isRawMessage(data)) {
-    key = EVENTS.RAW_MESSAGE
-    parsedData = data
-  } else if (JSONString) {
-    data = JSON.parse(data)
-    key = Object.keys(data)[0]
-    parsedData = data[key]
-  } else {
-    key = 'error'
-    parsedData = new Error(ERRORS.COULD_NOT_PARSE_MESSAGE)
+  const isBuffer = isBufferMessage(data)
+  const isJson = isJSONMessage(data)
+  const isString = isStringMessage(data)
+
+  // if (!data && isRaw) return { key: EVENTS.RAW_MESSAGE, data }
+
+  // // probably server-side
+  // if (!data) {
+  //   if (isRawMessage(data)) {
+  //     return { key: EVENTS.RAW_MESSAGE, data: data }
+  //   } else {
+  //     const json = JSON.parse(data as any)
+  //     const key = Object.keys(json)[0]
+  //     const value = Object.values(json)[0]
+  //     return { key: key, data: value }
+  //   }
+  // }
+
+  if (isJson) {
+    const object = JSON.parse(data)
+    const key = Object.keys(object)[0]
+    const value = object[key]
+    return { key: key, data: value }
   }
 
-  return { key: key, data: parsedData }
+  if (isBuffer) {
+    return { key: EVENTS.RAW_MESSAGE, data: data }
+  }
+
+  if (isString) {
+    return { key: EVENTS.RAW_MESSAGE, data: data }
+  }
+
+  return { key: 'error', data: new Error(ERRORS.COULD_NOT_PARSE_MESSAGE) }
 }
 
 export default ParseMessage
