@@ -1,31 +1,42 @@
-import { Data, RawMessage, EventName } from './types'
-import { isRawMessage } from './helpers'
-import { EVENTS } from './constants'
+import { Data, EventName, RawMessage } from './types.js'
+import { isBufferMessage, isStringMessage } from './helpers.js'
+import { EVENTS } from './constants.js'
 
 const SendMessage = (
-  dataChannel: RTCDataChannel,
+  dataChannel: any | RTCDataChannel,
   maxMessageSize: number | undefined,
   eventName: EventName,
   data: Data | RawMessage | null = null
 ) => {
-  const send = (data: any) => {
+  const send = (data: any, isBuffer: boolean) => {
     const bytes = data.byteLength ?? data.length * 2 // (times 2 for characters that uses 2 bytes per char)
 
     if (typeof maxMessageSize === 'number' && bytes > maxMessageSize) {
       throw new Error(`maxMessageSize of ${maxMessageSize} exceeded`)
     } else {
-      Promise.resolve().then(() => {
-        dataChannel.send(data)
-      })
+      Promise.resolve()
+        .then(() => {
+          // server-side (send() does not exist on the server side)
+          if (dataChannel.send) dataChannel.send(data)
+          else {
+            if (!isBuffer) dataChannel.sendMessage(data)
+            else dataChannel.sendMessageBinary(Buffer.from(data))
+          }
+        })
+        .catch(error => {
+          console.log('error', error)
+        })
     }
   }
 
-  if (dataChannel.readyState === 'open') {
+  if (!dataChannel) return
+
+  if (dataChannel.readyState === 'open' || dataChannel.isOpen?.()) {
     try {
-      if (eventName === EVENTS.RAW_MESSAGE && data !== null && isRawMessage(data)) {
-        send(data)
+      if (eventName === EVENTS.RAW_MESSAGE && data !== null && (isStringMessage(data) || isBufferMessage(data))) {
+        send(data, isBufferMessage(data))
       } else {
-        send(JSON.stringify({ [eventName]: data }))
+        send(JSON.stringify({ [eventName]: data }), false)
       }
     } catch (error) {
       console.error('Error in sendMessage.ts: ', error.message)
