@@ -3,30 +3,15 @@ import type { IncomingMessage, OutgoingMessage } from 'http'
 const ArrayBufferView = Object.getPrototypeOf(Object.getPrototypeOf(new Uint8Array())).constructor
 export { ArrayBufferView }
 
-export type USVString = string
 export type ChannelId = string | undefined
-export type EventName = string
-export type RoomId = ChannelId
 export type Data = string | number | Object
+export type EventName = string
 export type Payload = { [eventName: string]: Data }
 export type RawMessage = USVString | ArrayBuffer | ArrayBufferView
+export type RoomId = ChannelId
+export type USVString = string
 
 export interface ServerOptions {
-  iceServers?: RTCIceServer[]
-  iceTransportPolicy?: RTCIceTransportPolicy
-  label?: string
-  ordered?: boolean
-  maxRetransmits?: number
-  maxPacketLifeTime?: number
-  cors?: CorsOptions
-  autoManageBuffering?: boolean
-  /** Set a custom port range for the WebRTC connection. */
-  portRange?: {
-    /** Minimum port range (defaults to 0) */
-    min: number
-    /** Minimum port range (defaults to 65535) */
-    max: number
-  }
   /**
    * A async function to authenticate and authorize a user.
    * @param auth The authentication token
@@ -38,27 +23,45 @@ export interface ServerOptions {
     request: IncomingMessage,
     response: OutgoingMessage
   ) => Promise<boolean | any>
+  autoManageBuffering?: boolean
+  cors?: CorsOptions
+  iceServers?: RTCIceServer[]
+  /** @deprecated since version 2.0 (not available in node-datachannel) */
+  iceTransportPolicy?: RTCIceTransportPolicy
+  label?: string
+  /** @deprecated since version 2.0 (not available in node-datachannel) */
+  maxPacketLifeTime?: number
+  /** @deprecated since version 2.0 (not available in node-datachannel) */
+  maxRetransmits?: number
+  ordered?: boolean
+  /** Set a custom port range for the WebRTC connection. */
+  portRange?: {
+    /** Minimum port range (defaults to 0) */
+    min: number
+    /** Minimum port range (defaults to 65535) */
+    max: number
+  }
 }
 
 export interface ClientOptions {
-  iceServers?: RTCIceServer[]
-  iceTransportPolicy?: RTCIceTransportPolicy
-  url?: string
   authorization?: string | undefined
-  port?: number
+  iceServers?: RTCIceServer[] // eslint-disable-line no-undef
+  iceTransportPolicy?: RTCIceTransportPolicy // eslint-disable-line no-undef
   label?: string
+  port?: number
+  url?: string
 }
 
 export interface EmitOptions {
-  reliable?: boolean
   interval?: number
+  reliable?: boolean
   runs?: number
 }
 
 type CorsOptionsOriginFunction = (req: IncomingMessage) => string
 export interface CorsOptions {
-  origin: string | CorsOptionsOriginFunction
   allowAuthorization?: boolean
+  origin: string | CorsOptionsOriginFunction
 }
 
 export interface EventCallbackClient {
@@ -73,16 +76,21 @@ export interface EventCallbackRawMessage {
   (rawMessage: RawMessage): void
 }
 
+export interface ConnectionError extends Error {
+  status: number
+  statusText: string
+}
+
 export interface ConnectionEventCallbackClient {
-  (error?: Error): void
+  (error?: ConnectionError): void
 }
 
 export interface DisconnectEventCallbackServer {
-  (connectionState: 'disconnected' | 'failed' | 'closed'): void
+  (connectionState: 'closed' | 'disconnected' | 'failed'): void
 }
 
 export interface EventOptions {
+  id?: ChannelId
   roomId?: RoomId
   senderId?: ChannelId
-  id?: ChannelId
 }
